Add lazyFromValue function...
Once the Lazy type is based on a Tagged Union, it is then advantageous to have a `lazyFromValue` function added to the API so that when an immediately evaluated value that does not need to be deferred is desired, one can create it with this function.  The result is that one level of extra function call to `force` the evaluation is not required when this function is used.